start of early work to add opencontainers distribution
there is not much code in the core of the upstream (implemented
in golang) so I need to look closer to figure out how to
best implement this. I think ideally we would want some kind of
class to generate the correct data structure, although I am
not sure this will fit nicely into something like Django where
we would need to use database models and serialization instead.

Signed-off-by: vsoch <vsochat@stanford.edu>